feat(#TM-7): painting statistics from frontend

+++ server/internal/detail/detail.go
@@ -1,10 +1,10 @@
 package detail
 
 type Detail struct {
-	SellOffersMean         int     `json:"sellOffersMean"`
-	SellOffersStdDeviation float64 `json:"sellOffersStdDeviation"`
-	SellOffersMedian       int     `json:"sellOffersMedian"`
-	BuyOffersMean          int     `json:"buyOffersMean"`
-	BuyOffersStdDeviation  float64 `json:"buyOffersStdDeviation"`
-	BuyOffersMedian        int     `json:"buyOffersMedian"`
+	SellOffersMean         int `json:"sellOffersMean"`
+	SellOffersStdDeviation int `json:"sellOffersStdDeviation"`
+	SellOffersMedian       int `json:"sellOffersMedian"`
+	BuyOffersMean          int `json:"buyOffersMean"`
+	BuyOffersStdDeviation  int `json:"buyOffersStdDeviation"`
+	BuyOffersMedian        int `json:"buyOffersMedian"`
 }
